% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-data.R
\docType{data}
\name{binary-data}
\alias{binary-data}
\alias{binary_large}
\alias{binary_large-data}
\alias{binary_small}
\alias{binary_small-data}
\alias{binary_super}
\alias{binary_super-data}
\title{Sample simulated data for binary choice example in vignette}
\details{
The package provides four sample datasets for the
hierarchical binary choice model described in the vignette.
These datasets are: \describe{ \item{binary}{$N=50$, $k=3$}
\item{binary_small}{$N=20$, $k=2$} \item{binary_large}
{$N=800$, $k=3$} \item{binary_super}{$N=1200$, $k=3$} }
The datasets were generated using the code in data-raw/
binary_sim.R.
}

