% Generated by roxygen2 (4.0.2): do not edit by hand
\name{binary}
\alias{binary}
\alias{binary.f}
\alias{binary.grad}
\alias{binary.hess}
\title{Binary choice example}
\usage{
binary.f(P, Y, X, T, inv.Omega, inv.Sigma, order.row = FALSE)

binary.grad(P, Y, X, T, inv.Omega, inv.Sigma, order.row = FALSE)

binary.hess(P, Y, X, T, inv.Omega, inv.Sigma, order.row = FALSE)
}
\arguments{
  \item{P}{Numeric vector of length (N+1)*k.  First N*k
  elements are heterogeneous coefficients. The remaining k
  elements are population parameters.}

  \item{Y,X}{Data matrices}

  \item{inv.Omega,inv.Sigma}{Prior parameters}
}
\value{
Log posterior density, gradient and Hessian.
}
\description{
Functions for binary choice example in the vignette.
}
\details{
Hessian is sparse, and returned as a dgcMatrix object
}

