% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrices.R
\name{Coord.to.Pattern.Matrix}
\alias{Coord.to.Pattern.Matrix}
\title{Pattern matrix from row and column indices.}
\usage{
Coord.to.Pattern.Matrix(rows, cols, dims, compressed = TRUE,
  symmetric = FALSE)
}
\arguments{
  \item{rows,cols}{row and column indices of non-zero
  elements}

  \item{dims}{2-element vector for number of rows and
  columns in matrix}

  \item{compressed}{If TRUE, returns a matrix is compressed
  column (default=TRUE)}

  \item{symmetric}{If TRUE, matrix will be symmetric, and
  only the lower triangular elements need to be provided
  (default=FALSE)}
}
\value{
A sparse pattern matrix
}
\description{
Converts row and column indices to a pattern Matrix object
of Matrix class
}
\details{
This function is useful to prototype a sparsity pattern. No
assumptions are made about symmetry.
}

