% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrices.R
\name{Coord.to.Pointers}
\alias{Coord.to.Pointers}
\title{Column (row) indices and row (column) pointers from sparse matrix.}
\usage{
Coord.to.Pointers(rows, cols, dims, triangle = TRUE, lower = TRUE,
  order = c("column", "row", "symmetric"), index1 = TRUE)
}
\arguments{
  \item{rows,cols}{row and column indices of non-zero
  elements}

  \item{dims}{2-element vector for number of rows and
  columns in matrix}

  \item{triangle}{Is input intended to be a triangular
  (TRUE) or full (FALSE) matrix.  See Details for how this
  argument is interpreted for different values of
  \code{order}.}

  \item{lower}{If \code{triangular} is true, this argument
  identifies if the input matrix is lower- or
  upper-triangular.  This argument is ignored if
  \code{triangle} is FALSE.}

  \item{order}{Pointers can be ordered according to a
  row-oriented or column-oriented compression pattern.  If
  symmetric is chosen, the compression method does not
  matter.  The input matrix must be symmetric (although not
  necessarily stored that way) to get symmetric output.
  For symmetric input, the only difference between
  "symmetric," and either "row" or "column," is the names
  of the elements in the list. See details.}

  \item{index1}{TRUE if using 1-based indexing.  FALSE for
  0-based indexing.}
}
\value{
A list with two named elements.  See 'Matrix.to.Pointers'.
}
\description{
Returns list of column (row) indices row (column) pointers
of the non-zero elements of a sparse matrix stored in row
(column)-oriented compressed format.
}
\details{
\code{triangle} and \code{order} have the following
interpretation: \describe{ \item{triangle=TRUE}{Input
\code{rows} and {cols} represent lower or upper triangle of
a matrix. If \code{order="symmetric"}, then the output list
will be for a full, symmetric matrix. Otherwise, the output
list will be for only the lower or upper triangle.  Any
elements outside of the specified triangle will trigger an
error.} \item{triangle=FALSE}{Input \code{rows} and {cols}
represent a full matrix. If that matrix is not symmetric,
then \code{order=="symmetric"} will trigger an error.} }
}

