% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diffs.R
\name{get.indices}
\alias{get.indices}
\title{##}
\usage{
get.indices(k, Q)
}
\arguments{
  \item{k}{row or column for which you want the column or
  row indices of the non-zero elements}

  \item{x}{Value at which to evaluate Hessian ##}

  \item{df}{function that returns gradient at x ##}

  \item{W}{coloring scheme from color.cols function ##}

  \item{delta}{finite differencing factor ##}

  \item{a}{list representing the sparsity structure in
  column (row)-compressed format. The first element
  contains the row (column) indices, and the second
  contains the column (row) pointers.}
}
\value{
Finite difference gradients, with each group in a column.

An integer vector of row (column) indices.
}
\description{
Returns row (column) indices for non-zero elements in
column (row) k.
}
\details{
All indices are 1-based (as in R).  The Matrix.to.Pointers
function will return a list suitable for use in this
function.
}

