% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_colors}
\alias{get_colors}
\title{Vertex coloring of sparse symmetric matrix}
\usage{
get_colors(pntr, idx, nvars)
}
\arguments{
  \item{pntr,idx}{row pointers and column indices (CSC or
  CSR format; same since matrix is symmetric). Must use
  zero-based indexing.}

  \item{nvars}{Dimension of matrix (number of variables)}
}
\value{
A list.  Each element of the list represents a color, and
contains an integer vector with the indices of the
variables with that color.  Indices are zero-based.
}
\description{
Generate proper coloring of a sparse symmetric matrix.
}

