% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Matrix.to.Coord}
\alias{Matrix.to.Coord}
\title{Row and column indices from sparse symmetric matrix.}
\usage{
Matrix.to.Coord(M)
}
\arguments{
  \item{M}{A symmetric Matrix}
}
\value{
A list with two named elements. \describe{ \item{iRow}{
Integer vector containing row indices of non-zero elements
in the lower triangle of M} \item{jCol}{ Integer vector
containing column indices of non-zero elements in the lower
triangle of M} }
}
\description{
Returns list of row and column indices of the non-zero
elements of the lower triangle of a symmetric matrix.
}
\details{
Takes an object of class Matrix and returns the row and
column indices of the non-zero elements in the lower
triangle
}

