% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrices.R
\name{Matrix.to.Pointers}
\alias{Matrix.to.Pointers}
\title{Column (row) indices and row (column) pointers from sparse matrix.}
\usage{
Matrix.to.Pointers(M, order = c("column", "row"))
}
\arguments{
  \item{M}{A sparse Matrix, as defined in the Matrix
  package.}
}
\value{
A list with two named elements. \describe{ \item{rows}{
Integer vector containing row indices of non-zero elements}
\item{cols}{ Integer vector containing column indices of
non-zero elements} }
}
\description{
Returns list of column (row) indices row (column) pointers
of the non-zero elements of a sparse matrix stored in row
(column)-oriented compressed format.
}

