% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sparseHessianFD-class.R
\docType{class}
\name{sparseHessianFD-class}
\alias{sparseHessianFD}
\alias{sparseHessianFD-class}
\title{sparseHessianFD}
\description{
A reference class for computing sparse Hessians
}
\section{Fields}{

\describe{
\item{\code{fn1}}{A closure for calling fn(x, ...).}

\item{\code{gr1}}{A closure for calling gr(x, ...).}

\item{\code{iRow,jCol}}{Numeric vectors with row and column indices of the non-zero elements in the lower triangle (including diagonal) of the Hessian.}

\item{\code{eps}}{The perturbation amount for finite differencing of the gradient to compute the Hessian. Defaults to sqrt(.Machine$double.eps).}

\item{\code{index1}}{TRUE if rows and cols use 1-based (R format) indexing (FALSE for 0-based (C format) indexing.}

\item{\code{D}}{raw finite differences (internal use only)}

\item{\code{nvars}}{Number of variables (length of x)}

\item{\code{nnz}}{Number of non-zero elements in the lower triangle of the Hessian.}

\item{\code{ready}}{TRUE if object has been initialized, and Hessian has been partitioned.}

\item{\code{idx,pntr}}{Column indices and row pointers for non-zero elements in lower triangle of the Hessian.  Row-oriented compressed storage.}

\item{\code{colors}}{A list representation of the partitioning of the columns of the Hessian.  This is used as part of the estimation algorithm.  Specifically, each list element is a "color," and each element in the corresponding vector is a column with that color.  See references.}

\item{\code{colors_vec}}{A vector representation of the partitioning of the columns.  There are nvars elements, one for each column of the Hessian.  The value corresponds to the "color" for that column.}
}}
\section{Methods}{

\describe{
\item{\code{fn(x)}}{Return function value, evaluated at x: fn(x, ...)}

\item{\code{fngr(x)}}{Return list of function value and gradient, evaluated at x}

\item{\code{fngrhs(x)}}{Return list of function value, gradient, and Hessian, evaluated at x}

\item{\code{gr(x)}}{Return gradient, evaluated at x:  gr(x,...)}

\item{\code{hessian(x)}}{Return sparse Hessian, evaluated at x, as a dgCMatrix object.}

\item{\code{initialize(x, fn, gr, rows, cols, direct = NULL,
  eps = sqrt(.Machine$double.eps), index1 = TRUE, ...)}}{Initialize object with functions to compute the objective function and gradient (fn and gr), row and column indices of non-zero elements (rows and cols), an initial variable vector x at which fn and gr can be evaluated, a finite differencing parameter eps, flags for 0 or 1-based indexing (index1), whether sparsity pattern is just for the lower triangle (indexLT), and other arguments (...) to be passed to fn and gr.}

\item{\code{partition()}}{Return the partitioning used to compute finite differences}

\item{\code{pointers(out.index1 = index1)}}{Return list with indices (idx) and pointers (pntr) for sparsity pattern of the compressed sparse Hessian.  Since the Hessian is symmetric, the indices and pointers for row-oriented and column-oriented storage patterns are the same.}
}}

