% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sparseHessianFD-deprecated}
\alias{Coord.to.Sym.Pattern.Matrix}
\alias{Sym.CSC.to.Matrix}
\alias{new.sparse.hessian.obj}
\alias{sparseHessianFD-deprecated}
\title{Deprecated functions}
\usage{
Sym.CSC.to.Matrix(H, nvars)

Coord.to.Sym.Pattern.Matrix(H, nvars)

new.sparse.hessian.obj(x, fn, gr, hs, fd.method = 0L,
  eps = sqrt(.Machine$double.eps), ...)
}
\arguments{
  \item{H}{a list containing Hessian data.  See details.}

  \item{nvars}{the number of rows (and columns) in the
  matrix.}

  \item{x}{variable vector for initialization}

  \item{fn}{R function that returns function value}

  \item{gr}{R function that returns the gradient of the
  function}

  \item{hs}{list of two vectors: row and column indices of
  non-zero elements of lower triangle of Hessian.  See
  details.}

  \item{fd.method}{If TRUE, use direct method for
  computatation.  Otherwise, use indirect/substitution
  method.  See references.}

  \item{eps}{The perturbation amount for finite
  differencing of the gradient to compute the Hessian.
  Defaults to sqrt(.Machine$double.eps).}

  \item{...}{Other parameters to be passed to fn and gr.}
}
\value{
An object of Matrix class.
}
\description{
These functions were in earlier versions, but will no
longer be maintained, and are not even guaranteed to work
now.

Build sparse matrix from data in CSC (column compressed)
format.

Deprecated constructor
}
\details{
Use Matrix::sparseMatrix instead of Sym.CSC.to.Matrix.

Use Coord.to.Pattern.Matrix with symmetric=TRUE instead of
Coord.to.Sym.Pattern.Matrix.

hs is a list of two elements: \describe{ \item{iRow}{
Integer vector of row indices of non-zero elements in lower
triangle of Hessian.} \item{jCol}{ Integer vector of column
indices of non-zero elements in lower triangle of Hessian.}
}
}

