% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Sym.CSC.to.Matrix}
\alias{Sym.CSC.to.Matrix}
\title{Sym.CSC.to.Matrix}
\usage{
Sym.CSC.to.Matrix(H, nvars)
}
\arguments{
  \item{H}{a list containing Hessian data.  See details.}

  \item{nvars}{the number of rows (and columns) in the
  matrix.}
}
\value{
An object of Matrix class.
}
\description{
Build sparse matrix from data in CSC (column compressed)
format.
}
\details{
H is a list with three elements. \describe{ \item{iRow}{Row
indices for each of the non-zero elements in the lower
triangle of H} \item{jpntr}{A vector of length
\code{nvars+1}.  jpntr[j] is the index of the element in
vals that is the first non-zero element in the jth column
of the matrix.} \item{vals}{The values of the non-zero
elements in the lower triangle of the matrix.} } The input
list contains information about only the non-zero elements
in the lower triangle of the matrix.
}

