% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-data.R
\docType{data}
\name{binary-data}
\alias{binary-data}
\alias{binary_small}
\alias{binary_large}
\alias{binary_super}
\alias{binary_small-data}
\alias{binary_large-data}
\alias{binary_super-data}
\title{Sample simulated data for binary choice example in vignette}
\description{
Sample datasets for vignette
}
\details{
The package provides four sample datasets for the hierarchical binary choice model described in the vignette. These datasets are:
\describe{
\item{binary}{\eqn{N=50}, \eqn{k=3}}
\item{binary_small}{\eqn{N=20}, \eqn{k=2}}
\item{binary_large}{\eqn{N=800}, \eqn{k=3}}
\item{binary_super}{\eqn{N=1200}, \eqn{k=3}}
}
\eqn{N} is the number of heterogeneous units.
\eqn{k} is the number of covariates.

The datasets were generated using the code in data-raw/binary_sim.R.
}
