% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_colors}
\alias{get_colors}
\title{Vertex coloring of a sparse undirected graph}
\usage{
get_colors(pntr, idx, nvars)
}
\arguments{
\item{pntr, idx}{row pointers and column indices of the adjacency matrix, in compressed column-oriented format. Must use zero-based indexing.}

\item{nvars}{Number of vertices.}
}
\value{
An integer vector of length nvars, where each element represents the color of the corresponding vertex. Indices are zero-based.
}
\description{
Generate proper vertex coloring of a sparse undirected graph.
}
\details{
For internal use.  You should not have to call this function directly.
}
