% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrices.R
\name{Matrix.to.Coord}
\alias{Matrix.to.Coord}
\title{Row and column indices from sparse matrix.}
\usage{
Matrix.to.Coord(M, index1 = TRUE)
}
\arguments{
  \item{M}{A matrix that can be coerced into a
  \code{TsparseMatrix} object, as defined in the Matrix
  package.  Base R matrices, and most Matrix matrices, are
  valid.}

  \item{index1}{TRUE if the index of the first element
  should be 1, and FALSE if 0.}
}
\value{
A list with two named elements. \describe{ \item{rows}{
Integer vector containing row indices of non-zero elements}
\item{cols}{ Integer vector containing column indices of
non-zero elements} }
}
\description{
Returns list of row and column indices of the non-zero
elements of a sparse matrix.
}

