% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrices.R
\name{Matrix.to.Coord}
\alias{Matrix.to.Coord}
\title{Row and column indices from sparse matrix.}
\usage{
Matrix.to.Coord(M)
}
\arguments{
  \item{M}{A sparse Matrix, as defined in the Matrix
  package.}
}
\value{
A list with two named elements. \describe{ \item{rows}{
Integer vector containing row indices of non-zero elements}
\item{cols}{ Integer vector containing column indices of
non-zero elements} }
}
\description{
Returns list of row and column indices of the non-zero
elements of a sparse matrix.
}

