% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subst}
\alias{subst}
\title{Estimate sparse Hessian}
\usage{
subst(Y, colors, W, jCol, ipntr, delta, nvars, nnz)
}
\arguments{
  \item{Y}{Matrix of finite differences}

  \item{colors}{Vector of length nvars that identifies
  color of each variable}

  \item{W}{A list.  Each element represents a color, and
  contains an integer vector with the indices of the
  variables with that color.  Indexing is zero-based.}

  \item{jCol,ipntr}{Column indices and row pointers for
  non-zero elements of lower triangle of Hessian.}

  \item{delta}{Perturbation factor used to compute finite
  differences.}

  \item{nvars}{Dimension of Hessian (number of variables)}

  \item{nnz}{Number of non-zero elements in the lower
  triangle of the Hessian.}
}
\value{
A sparse Hessian of class dgCMatrix.
}
\description{
Estimate Hessian using triangular subsitution algorithm
}

