% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subst}
\alias{subst}
\title{Estimate sparse Hessian}
\usage{
subst(Y, colors, jCol, ipntr, delta, nvars, nnz)
}
\arguments{
\item{Y}{Matrix of finite differences of gradients}

\item{colors}{Vector of length nvars that identifies color of each variable}

\item{jCol, ipntr}{Column indices and row pointers for non-zero elements of lower triangle of Hessian (row-oriented compressed format).}

\item{delta}{Perturbation factor used to compute finite differences of gradients.}

\item{nvars}{Dimension of Hessian (number of variables)}

\item{nnz}{Number of non-zero elements in the lower triangle of the Hessian.}
}
\value{
A sparse Hessian of class dgCMatrix.
}
\description{
Estimate Hessian using triangular subsitution algorithm
}
\details{
For internal use.  You should not have to call this function directly.
}
